fix(financeiro): fechamento não carimba conta rateada sem gerar fatura
Três defeitos que se encobriam. Sintoma relatado: o síndico criava uma conta de
consumo, rodava o fechamento, a conta aparecia como "rateada" e o morador não
via cobrança nenhuma — sem erro em lugar algum.

1. O fim do fechamento marcava TODA conta como RATEADA, mesmo tendo criado zero
   faturas. Agora só carimba se alguma fatura saiu.

2. Conta em modo FRACAO_IDEAL sem fração cadastrada fazia
   `Math.floor(valor * 0 / 0)` = NaN, e `NaN > 0` é falso — o item era
   descartado em silêncio. O fechamento agora recusa antes de emitir e diz quais
   contas dependem de fração que não existe. Guarda extra no cálculo para um
   caminho novo não reintroduzir o sumiço.

3. SUM() sobre coluna BIGINT devolve NUMERIC, que o driver entrega como string,
   e o parser de tipos só converte BIGINT. Os KPIs vinham com `totalCentavos`
   em texto: uma subtração funcionava por coerção, mas uma soma concatenava e o
   total virava lixo sem erro. Resolvido com ::bigint na consulta.

+++ services/financeiro/models/faturasModel.js
@@ -159,9 +159,13 @@ export async function kpisPorCompetencia(condominioId, competencia) {
   const { rows } = await consultar(
     `SELECT
        COUNT(*) FILTER (WHERE status <> 'CANCELADA') AS total,
-       COALESCE(SUM(valor_centavos) FILTER (WHERE status <> 'CANCELADA'), 0) AS "totalCentavos",
+       -- ::bigint nao e enfeite: SUM() sobre BIGINT devolve NUMERIC, que o
+       -- driver entrega como string, e o parser de tipos so converte BIGINT.
+       -- Sem o cast, uma subtracao funciona por coercao, mas uma soma
+       -- concatena as strings e o total vira lixo sem erro nenhum.
+       COALESCE(SUM(valor_centavos) FILTER (WHERE status <> 'CANCELADA'), 0)::bigint AS "totalCentavos",
        COUNT(*) FILTER (WHERE status = 'PAGA') AS pagas,
-       COALESCE(SUM(valor_centavos) FILTER (WHERE status = 'PAGA'), 0) AS "pagasCentavos",
+       COALESCE(SUM(valor_centavos) FILTER (WHERE status = 'PAGA'), 0)::bigint AS "pagasCentavos",
        COUNT(*) FILTER (WHERE status IN ('ABERTA', 'EM_ATRASO')) AS abertas,
        COUNT(*) FILTER (WHERE status = 'EM_ATRASO') AS "emAtraso"
      FROM faturas
